requirements.txt
ccxt==4.1.87
loguru==0.7.2
pandas==2.1.4
tqdm==4.66.1
web3==6.11.4
requests==2.31.0
openpyxl==3.1.2
msoffcrypto-tool==5.1.1